add delete confirm step.

+++ lib/widgets/tasks_list.dart
@@ -3,6 +3,37 @@ import 'package:todoey_flutter/widgets/task_tile.dart';
 import 'package:provider/provider.dart';
 import 'package:todoey_flutter/models/task_data.dart';
 
+enum ConfirmAction { CANCEL, ACCEPT}
+Future<ConfirmAction> _asyncConfirmDialog(BuildContext context) async {
+  return showDialog<ConfirmAction>(
+    context: context,
+    barrierDismissible: false, // user must tap button for close dialog!
+    builder: (BuildContext context) {
+      return AlertDialog(
+        title: Text('Do you want delete task?'),
+        content: const Text(
+            'This will delete your task from list.'),
+        actions: <Widget>[
+          FlatButton(
+            child: const Text('CANCEL'),
+            onPressed: () {
+              Navigator.of(context).pop(ConfirmAction.CANCEL);
+            },
+          ),
+          FlatButton(
+            child: const Text('ACCEPT'),
+            onPressed: () {
+              Navigator.of(context).pop(ConfirmAction.ACCEPT);
+
+            },
+          )
+        ],
+      );
+    },
+  );
+}
+
+
 class TasksList extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
@@ -17,8 +48,13 @@ class TasksList extends StatelessWidget {
               checkboxCallback: (checkboxState) {
                 taskData.updateTask(task);
               },
-              longPressCallback: () {
-                taskData.deleteTask(task);
+              longPressCallback: () async {
+//                taskData.deleteTask(task);
+                final ConfirmAction action = await _asyncConfirmDialog(context);
+                print('Confirm Action $action');
+                if(action == ConfirmAction.ACCEPT) {
+                  taskData.deleteTask(task);
+                }
               },
             );
           },
@@ -27,4 +63,6 @@ class TasksList extends StatelessWidget {
       },
     );
   }
+
+
 }
